# src/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Project 1 API",
    "description": "Send forward your gaming community using Project 1 API",
    "version": "1.0.0"
  },
  "produces": [
    "application/json"
  ],
...
  "basePath": "/api",
  "paths": {
    "/auth": {
      "post": {
        "x-swagger-router-controller": "AuthController",
        "operationId": "authenticate",
        "tags": [
          "Security"
        ],
        "summary": "Authenticates a user",
        "description": "Authenticates the user against the DB.",
        "parameters": [
          {
            "name": "authRequest",
            "in": "body",
            "description": "Credentials object to authenticate",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User has been successfully logged in"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/games": {
      "get": {
        "x-swagger-router-controller": "GameController",
        "operationId": "getGames",
        "tags": [
          "Game"
        ],
        "summary": "Gets all games",
        "description": "Gets `Game` objects.\nOptional query param of **size** determines\nsize of returned array\n",
        "parameters": [
          {
            "name": "size",
            "in": "query",
            "description": "Size of array",
            "required": false,
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of Games",
            "schema": {
              "title": "ArrayOfGames",
              "type": "array",
              "items": {
                "$ref": "#/definitions/Game"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/games/addGame": {
      "post": {
        "x-swagger-router-controller": "GameController",
        "operationId": "addGame",
        "tags": [
          "Game"
        ],
        "summary": "Adds a game",
        "description": "Adds new `Game` object to the DB.",
        "parameters": [
          {
            "name": "game",
            "in": "body",
            "description": "Game object to insert",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Game"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game has been successfully added to the DB"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/games/{gameId}/getGameById": {
      "get": {
        "x-swagger-router-controller": "GameController",
        "operationId": "getGameById",
        "tags": [
          "Game"
        ],
        "summary": "Get the specified game",
        "description": "Gets `Game` object specified by its ID.",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "gameId of Game object to search for.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved Game object",
            "schema": {
              "$ref": "#/definitions/Game"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/games/{gameName}/getGameByName": {
      "get": {
        "x-swagger-router-controller": "GameController",
        "operationId": "getGameByName",
        "tags": [
          "Game"
        ],
        "summary": "Get the corresponding game based on the game name",
        "description": "Gets `Game` object specified by its name.",
        "parameters": [
          {
            "name": "gameName",
            "in": "path",
            "description": "Game's name to search for.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved Game object",
            "schema": {
              "$ref": "#/definitions/Game"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/games/{gameId}/updateGame": {
      "put": {
        "x-swagger-router-controller": "GameController",
        "operationId": "updateGame",
        "tags": [
          "Game"
        ],
        "summary": "Update the corresponding game",
        "description": "Update `Game` object specified by its ID.",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "gameId of Game object to update.",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameObject",
            "in": "body",
            "description": "Game object that will update the old one.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Game"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated Game object",
            "schema": {
              "$ref": "#/definitions/Game"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/games/{gameId}/deleteGame": {
      "put": {
        "x-swagger-router-controller": "GameController",
        "operationId": "deleteGame",
        "tags": [
          "Game"
        ],
        "summary": "Deactivate the corresponding game",
        "description": "Deactivates the 'Game' object specified by its ID.",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "gameId of Game object to deactivate.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated Game object",
            "schema": {
              "$ref": "#/definitions/Game"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AuthRequest": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "format": "password"
        }
      }
    },
...
    "Game": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "releaseDate": {
          "type": "string",
          "format": "date"
        },
        "multiPlayer": {
          "type": "boolean"
        },
        "editor": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date"
        },
        "updated_at": {
          "type": "string",
          "format": "date"
        }
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    }
  }
}